feat(theme): setup react-native-paper, AppContext theme
