+++ .env
@@ -1,0 +1,1 @@
+API_BASE_URL=https://jsonplaceholder.typicode.com
