+++ .gitignore
@@ -74,5 +74,6 @@ yarn-error.log
 !.yarn/sdks
 !.yarn/versions
 
-.env*
+.env*/
 /._stash
+/src/screens/wip
