--- src/components/organisms/home/index.ts
@@ -1,1 +1,0 @@
-export {default as Section} from './Section';
