+++ src/screens/main/HomeScreen/styles.ts
@@ -1,10 +1,11 @@
 import {StyleSheet} from 'react-native';
 import {Colors} from 'react-native/Libraries/NewAppScreen';
+import {useTheme} from 'react-native-paper';
 
 /*
  * To keep the template simple and small we're adding padding to prevent view
  * from rendering under the System UI.
- * For bigger apps the reccomendation is to use `react-native-safe-area-context`:
+ * For bigger apps the recommendation is to use `react-native-safe-area-context`:
  * https://github.com/AppAndFlow/react-native-safe-area-context
  *
  * You can read more about it here:
@@ -12,28 +13,22 @@ import {Colors} from 'react-native/Libraries/NewAppScreen';
  */
 const SAFE_PADDING = '5%';
 
-type ParamsType = {
-  isDarkMode?: boolean;
-};
+type PropsType = {muted?: boolean};
 
-export const homeStyles = {
-  static: StyleSheet.create({
-    header: {
-      paddingRight: SAFE_PADDING,
+export function homeScreenStyles(props?: PropsType) {
+  const theme = useTheme();
+  return StyleSheet.create({
+    _: {},
+    container: {
+      backgroundColor: theme.dark ? Colors.darker : Colors.lighter,
     },
-  }),
-  dynamic: ({isDarkMode}: ParamsType) =>
-    StyleSheet.create({
-      container: {
-        backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
-      },
-      content: {
-        backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
-      },
-      main: {
-        backgroundColor: isDarkMode ? Colors.black : Colors.white,
-        paddingHorizontal: SAFE_PADDING,
-        paddingBottom: SAFE_PADDING,
-      },
-    }),
-};
+    content: {
+      backgroundColor: theme.dark ? Colors.darker : Colors.lighter,
+    },
+    main: {
+      backgroundColor: theme.dark ? Colors.black : Colors.white,
+      paddingHorizontal: SAFE_PADDING,
+      paddingBottom: SAFE_PADDING,
+    },
+  });
+}
